07_jwt/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib
python-multipart
pydantic
bcrypt